add resizing of E and some calculations of BeamW
The user can now enter a maximum energy in (eV) and the program will
crop E and dE accordingly.

The script also now calculates and returns a BeamWeight_discrete, which
calculates the beam weights in a discrete manner instead of continuous.
This is of use in some parts of the program.

The way θ_to_BeamWeight is calculated is also changed. It is now
properly normalised so that we get 1 when summing along each beam.