fix a very nasty bug
If the calculation of a spectra of secondaries or primaries is called, it
checks for the presence of cascading matrices. If it does find any, it triggers
a calculation of the cascading matrix.

In the case of O, the spectras are calculated with the help of some parameters,
one of them being called A. Empirically, it has to be multiplied by 1.25 when
calculating the spectra, but not when calculating the cascading matrix.
The bug was that if a calculation of the cascading matrix was triggered, then A
would take the non-modified value to calculate the matrix, and then that value
would also be used for calculating the spectra (instead of the modified by 1.25
value). All consequent calls to the function would use the correct A*1.25 value
as the cascading matrix was already calculated, so that affected only the
highest energy. But because it is the first and highest energy, it did trickle
down the whole chain of energies. It is worth nothing though that it did not
modify the values by much, and that this would happen only when the cascading
matrix for O had to be computed.